Added the option of a 'Natural' boundary condition
This bounds works by extrapolating the boundary coefficients twice instead of once, then interpolating inside the 'inner boundary', and finally linear interpolation at the 'outer boundary' and out to infinity. This creates a natural transition from the domain to the linear extrapolation, and of as high continuity as the kernel itself.